Updated authController.js
Created login function. It checks if the user exists in the database by email by using the findOne Mongoose function. Then it verifies the password
